refactor(native): remove OpenXR-hook TU + un-build SA session, preserve quilt seed (#166)
Provider-only refactor, native side. The custom IUnityXRDisplay provider
(displayxr_xrprovider/) is the shipping path; the legacy xrGetInstanceProcAddr
hook chain is removed.

- Delete displayxr_hooks.cpp. Re-home the symbols surviving TUs still need
  (logger, real-fn-pointer storage — inert in provider mode, win32 window-binding
  injector, viewport/canvas-rect accessors) into new displayxr_native_shared.cpp.
- Rename displayxr_hooks.h -> displayxr_exports.h (public P/Invoke ABI) and
  displayxr_hooks_internal.h -> displayxr_backend.h (GraphicsBackend + backend glue);
  drop the hook-install + SA-only accessor decls.
- Drop displayxr_hooks.cpp + all displayxr_standalone*.{cpp,m} from CMake SOURCES.
  The SA render-to-atlas core stays on disk (dormant) as the seed for the future
  many-view quilt path (ADR-007) — banner added to displayxr_standalone.cpp.

DLL links clean without the hook/SA TUs; provider export surface intact.
