+++ native~/displayxr_local2d.cpp
@@ -14,7 +14,7 @@
 
 #include "displayxr_extensions.h"
 #include "displayxr_shared_state.h"
-#include "displayxr_hooks_internal.h"
+#include "displayxr_backend.h"
 
 #include <stdio.h>
 #include <string.h>
